add code which uses plain numpy plus joblib
this runs for our test datasets with 12x24 horizontal grid points
in 20 seconds on one core, all grid points and all days of year.
with joblib on 3 cores, it runs in 13.2 seconds.
We quickly squeezed data into 365-day years to fit into standard
numpy array. This needs to be improved, specifically:

* monthly GMT is interpolated to daily in a very simple way so far.
* treatment of 366-day years is very simple.