added a check for number of chains in loaded trace
this check catches traces, that have been saved partly (not evry chain),
by comparing the number of chains (nchains) in a loaded trace, with the
specified number of chains in the instance of the estimator class.
Saving of some chains of a sample happens, when the process is stopped
while sampling the second or a later chain.