New scaling method
Since we are not transforming data anymore, the scaling functions are
now reduced to one (might need a double check for correctness). It takes
into account, the threshold values, taken from a dict (threshold) and
also has an option to be called with "gmt=True", that disables calling
of threshold values, to scale gmt values.