cleanup and parameter changes
played around with some of the parameters for the priors. Still only works properly for Beta, Gamma and Normal. This still needs a lot of work and a way to pass the right parameters conveniently when switching variables. Most common error is: "Bad initial energy", mostly caused by bad estimation of observed variable. Another issue might be the way the model is passed to the "observed" pymc3 distribution (for Rice: nu="model", etc.)