implement exp transform of gmt_param in Beta model
this is needed to avoid negative values in the parameters,
which lead to errors.
We cannot split the sum of parameters like before as they are in
log(..).
We therefore use the full estimated parameter that feeds into
the Beta model, and create a reference value for every day of the year.
This relies on pandas datetime functionality.