Allow different models for variance (sigma).
`full` model has gmt driven yearly mean trend and
a shift in the yearly cycle through GMT.

`no_gmt_cycle_trend` is the full model without the shift
in the yearly cycle

`no_gmt_trend` excludes trends through GMT completely
and assumes a stationary sigma over the years. A yearly
cycle is included, but it is the same for each year.