Removed gmt - gmtv inconsistencies. dropna is done within the model now
the get_valid_subset function was used for 2 things:
1. get a random subset of the data
2. drop nan values (i.e. dry days for precip)
The second part made it impossible to use this data to learn the pbern parameter for precip, as this parameter needs the dry days. For the parametrization of precip-amounts on the other hands, nan values need to be droped.
Before this commit there was a bug in the precip model, as it used not only the subsampled data but the whole dataset to fit pbern.
Now drop nan values is done in the model setup functions directly, where it is needed.
For clarity, I renamed df_valid to df_subset