Add yearly sigma cycle, issues remain
If we follow our longerm model and define
a_sigma = pm.Exponential - b_sigma - yearly_sigma

We get the problem that a_sigma depends on time.
We can keep it out of the upper equation, but then it is not
secured that sigma can be close to zero for temperature close to 1.

Also: sampling is very slow for the full timeseries (40000 points).
This may be amended through adjusted priors.