First implementation of TasCycle with Relu transform
Relu(x) = rectified_linear_unit(x) = max(0,x)

the parameter models are linear models. Sigma is then transformed with the ReLU function. That insures the correct range of sigma (all negative values are set to zero). This is not the optimal prior for the sigma value but might work just well.