Reworked postprocessing
Now we can write netcdf with `write_netcdf.py`
as an alternative to `merge_cfact.py`

This seems to be of similar speed for small files
and addresses #71.

We take all metadata from the source file and we
use better variable names.

Only variables that are in `report_to_netcdf`
of settings.py are written.